==> scripts/03_eda.Rmd <==
---
title: "Exploratory Data Analysis"
output:
  html_document:
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(tidyr)
library(here)
library(gridExtra)

load(file = here("data", "all_sessions.Rda"))
load(file = here("data", "all_businesses.Rda"))
load(file = here("data", "all_businesses_eda.Rda"))
load(file = here("data", "eda.Rda"))
load(file = here("data", "all_MemberCouncil.Rda"))
load(file = here("data", "all_businesses_party.Rda"))
load(file = here("data", "all_businesses_wc.Rda"))
```

```{r message=FALSE, warning=FALSE, include=FALSE}
# Set timeframe from March 2004 to December 2023
all_businesses_eda <- eda %>% left_join(dplyr::select(all_businesses_eda, -TagNames, -SubmissionSession), by = "BusinessShortNumber")
all_businesses_party <- eda %>% left_join(dplyr::select(all_businesses_party, -PartyAbbreviation), by = "BusinessShortNumber")
```


## Counts
```{r echo=FALSE, message=FALSE, warning=FALSE}
n_all_environmental <- all_businesses_eda %>% dplyr::filter(grepl("Umwelt", TagNames))

n_sp_all <- all_businesses_party %>% dplyr::filter(grepl("SP", PartyAbbreviation))
n_sp_environmental <- n_all_environmental %>% inner_join(n_sp_all, by = "BusinessShortNumber")

n_gps_all <- all_businesses_party %>% dplyr::filter(grepl("GRÜNE", PartyAbbreviation))
n_gps_environmental <- n_all_environmental %>% inner_join(n_gps_all, by = "BusinessShortNumber")

n_all <- nrow(all_businesses_eda)
n_sp_all <- nrow(n_sp_all)
n_gps_all <- nrow(n_gps_all)
n_all_environmental <- nrow(n_all_environmental)
n_sp_environmental <- nrow(n_sp_environmental)
n_gps_environmental <- nrow(n_gps_environmental)
```

|             | All Bills       | Environmental Bills   |
|:------------|:----------------|:----------------------|
| All Parties | `r n_all`       | `r n_all_environmental` |
| SP          | `r n_sp_all`    | `r n_sp_environmental`  |
| Green Party | `r n_gps_all`   | `r n_gps_environmental` |



## Tags Available in Parliamentary Data

```{r echo=FALSE, message=FALSE, warning=FALSE}

all_tags <- all_businesses %>% distinct(TagNames) %>% drop_na()

# Assuming your dataframe is named df
tags_list <- strsplit(all_tags$TagNames, split = "\\|") # Step 1: Split the strings
tags_vector <- unlist(tags_list) # Step 2: Unlist to get a vector
unique_tags <- unique(tags_vector) # Step 3: Extract unique tags
unique_tags
```


## Number of Bills over Time
Dotted black vertical lines mark parliamentary elections.  
Dashed red vertical lines mark crises: The Fukushima nuclear accident (2011) and Greta Thunberg's climate strike (2018).
```{r echo=FALSE, message=FALSE, warning=FALSE}
eda_all_bills <- eda %>% 
  group_by(year_session) %>% 
  count() %>% 
  dplyr::rename("count_all_bills" = "n")

eda_sp_bills <- eda %>% 
  dplyr::filter(grepl("^SP$", PartyAbbreviation)) %>% 
  group_by(year_session) %>% 
  count() %>% 
  dplyr::rename("count_sp_bills" = "n")

eda_gps_bills <- eda %>% 
  dplyr::filter(grepl("^GRÜNE$", PartyAbbreviation)) %>% 
  group_by(year_session) %>% 
  count() %>% 
  dplyr::rename("count_gps_bills" = "n")

eda_plot <- eda_all_bills %>% 
  left_join(eda_sp_bills, by = "year_session") %>% 
  left_join(eda_gps_bills, by = "year_session") %>% 
  pivot_longer(cols = c(count_all_bills, count_sp_bills, count_gps_bills), names_to = "bill_type", values_to = "count")

ggplot(eda_plot, aes(x = year_session, y = count, color = bill_type)) +
  geom_line() + geom_point() + 
  scale_color_manual(values = c("count_gps_bills" = "springgreen3", "count_sp_bills" = "red", "count_all_bills" = "black"), 
                     labels = c("count_gps_bills" = "Green Party", "count_sp_bills" = "SP", "count_all_bills" = "All Bills")) +
  labs(title = "Number of Bills Introduced by MPs or Parties", x = " ", y = " ", color = " ") +
  # Mark elections
  geom_vline(xintercept = as.Date("2019-10-20"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = as.Date("2015-10-18"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = as.Date("2011-10-23"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = as.Date("2007-10-21"), linetype = "dotted", color = "#333333", size = 0.25) +
  # Mark crisis
  geom_vline(xintercept = as.Date("2011-03-11"), linetype = "dashed", color = "red", size = 0.25) +
  geom_vline(xintercept = as.Date("2018-08-20"), linetype = "dashed", color = "red", size = 0.25) +
  theme_minimal() +
  theme(legend.position = "bottom", legend.direction = "horizontal", panel.grid.minor = element_blank())
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
eda_environment_bills <- eda %>% 
  dplyr::filter(grepl("Umwelt", TagNames)) %>% 
  group_by(year_session) %>% 
  count() %>% 
  dplyr::rename("count_environment_bills" = "n")

eda_plot <- eda_all_bills %>% 
  left_join(eda_environment_bills, by = "year_session") %>% 
  pivot_longer(cols = c(count_all_bills, count_environment_bills), names_to = "bill_type", values_to = "count")

ggplot(eda_plot, aes(x = year_session, y = count, group = bill_type, color = as.factor(bill_type))) +
  geom_line() + geom_point() + 
  scale_color_manual(values = c("count_environment_bills" = "springgreen3", "count_all_bills" = "black"), 
                     labels = c("count_environment_bills" = "Environmental Bills", "count_all_bills" = "All Bills")) +
  labs(title = "Number of Bills Introduced by MPs or Parties", x = " ", y = " ", color = " ") +
  # Mark elections
  geom_vline(xintercept = as.Date("2019-10-20"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = as.Date("2015-10-18"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = as.Date("2011-10-23"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = as.Date("2007-10-21"), linetype = "dotted", color = "#333333", size = 0.25) +
  # Mark crisis
  geom_vline(xintercept = as.Date("2011-03-11"), linetype = "dashed", color = "red", size = 0.25) +
  geom_vline(xintercept = as.Date("2018-08-20"), linetype = "dashed", color = "red", size = 0.25) +
  theme_minimal() +
  theme(legend.position = "bottom", legend.direction = "horizontal", panel.grid.minor = element_blank())
```

```{r echo=FALSE, message=FALSE, warning=FALSE}

set_tag <- "Umwelt"
# set_tag <- "Landwirtschaft"
# set_tag <- "Energie"

eda_environment_bills <- eda %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  group_by(year_session) %>% 
  count() %>% 
  dplyr::rename("count_environment_bills" = "n")

eda_environment_sp_bills <- eda %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("^SP$", PartyAbbreviation)) %>% 
  group_by(year_session) %>% 
  count() %>% 
  dplyr::rename("count_environment_sp_bills" = "n")

eda_environment_gps_bills <- eda %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("^GRÜNE$", PartyAbbreviation)) %>% 
  group_by(year_session) %>% 
  count() %>% 
  dplyr::rename("count_environment_gps_bills" = "n")

eda_environment_svp_bills <- eda %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("^SVP$", PartyAbbreviation)) %>% 
  group_by(year_session) %>% 
  count() %>% 
  dplyr::rename("count_environment_svp_bills" = "n")

eda_environment_mitte_bills <- eda %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("^M-E$|^CVP$", PartyAbbreviation)) %>% 
  group_by(year_session) %>% 
  count() %>% 
  dplyr::rename("count_environment_mitte_bills" = "n")

eda_environment_fdp_bills <- eda %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("FDP", PartyAbbreviation)) %>% 
  group_by(year_session) %>% 
  count() %>% 
  dplyr::rename("count_environment_fdp_bills" = "n")

eda_environment_glp_bills <- eda %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("glp", PartyAbbreviation)) %>% 
  group_by(year_session) %>% 
  count() %>% 
  dplyr::rename("count_environment_glp_bills" = "n")

eda_plot <- eda_environment_bills %>% 
  left_join(eda_environment_sp_bills, by = "year_session") %>% 
  left_join(eda_environment_gps_bills, by = "year_session") %>% 
  left_join(eda_environment_svp_bills, by = "year_session") %>% 
  left_join(eda_environment_mitte_bills, by = "year_session") %>% 
  left_join(eda_environment_fdp_bills, by = "year_session") %>% 
  left_join(eda_environment_glp_bills, by = "year_session") %>% 
  pivot_longer(cols = c(
    count_environment_bills,
    count_environment_sp_bills,
    count_environment_gps_bills,
    count_environment_svp_bills,
    count_environment_mitte_bills,
    count_environment_glp_bills,
    count_environment_fdp_bills
    ), names_to = "bill_type", values_to = "count")

ggplot(eda_plot, aes(x = year_session, y = count, group = bill_type, color = as.factor(bill_type))) +
  geom_line() + geom_point() + 
  scale_color_manual(values = c("count_environment_bills" = "black",
                                "count_environment_sp_bills" = "red",
                                "count_environment_svp_bills" = "light green",
                                "count_environment_mitte_bills" = "orange",
                                "count_environment_fdp_bills" = "blue",
                                "count_environment_glp_bills" = "aquamarine4",
                                "count_environment_gps_bills" = "springgreen3"), 
                     labels = c("count_environment_bills" = "Environmental Bills",
                                "count_environment_sp_bills" = "SP",
                                "count_environment_svp_bills" = "SVP",
                                "count_environment_mitte_bills" = "Mitte",
                                "count_environment_fdp_bills" = "FDP",
                                "count_environment_glp_bills" = "GLP",
                                "count_environment_gps_bills" = "Green Party")) +
  labs(title = "Number of Environmental Bills Introduced by MPs or Parties", x = " ", y = " ", color = " ") +
  # Mark elections
  geom_vline(xintercept = as.Date("2019-10-20"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = as.Date("2015-10-18"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = as.Date("2011-10-23"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = as.Date("2007-10-21"), linetype = "dotted", color = "#333333", size = 0.25) +
  # Mark crisis
  geom_vline(xintercept = as.Date("2011-03-11"), linetype = "dashed", color = "red", size = 0.25) +
  geom_vline(xintercept = as.Date("2018-08-20"), linetype = "dashed", color = "red", size = 0.25) +
  theme_minimal() +
  theme(legend.position = "bottom", legend.direction = "horizontal", panel.grid.minor = element_blank())
```


```{r eval=FALSE, message=FALSE, warning=FALSE, include=FALSE}

set_tag <- "Umwelt"

ch_parties$LegislativePeriodNumber <- 47:52

ch_parties_long <- ch_parties %>% 
  dplyr::select(-year) %>% 
  # wide to long
  pivot_longer(!LegislativePeriodNumber, names_to = "PartyAbbreviation", values_to = "VoteShare") %>% 
  mutate(PartyAbbreviation = case_when(PartyAbbreviation == "sp" ~ "SP",
                                       PartyAbbreviation == "gps" ~ "GRÜNE",
                                       PartyAbbreviation == "mitte" ~ "M-E",
                                       T ~ PartyAbbreviation)) %>% 
  mutate(PartyAbbreviation = toupper(PartyAbbreviation))

session_leg_index <- all_sessions %>% 
  dplyr::select(ID, LegislativePeriodNumber) %>% 
  distinct(ID, .keep_all = T)

eda_environment_sp_bills_vote_share <- eda %>% 
  dplyr::select(-StartDate, -year_session) %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("^SP$", PartyAbbreviation)) %>% 
  left_join(session_leg_index, by = c("SubmissionSession" = "ID")) %>% 
  # group_by(PartyAbbreviation, LegislativePeriodNumber) %>% 
  group_by(PartyAbbreviation, SubmissionSession) %>% 
  count() %>% 
  mutate(LegislativePeriodNumber = as.numeric(substr(SubmissionSession, 1, 2))) %>%
  left_join(ch_parties_long, by = c("LegislativePeriodNumber", "PartyAbbreviation")) %>%
  dplyr::rename("count_environment_bills" = "n") %>% 
  mutate(bills_per_vote_share = count_environment_bills / VoteShare)

eda_environment_gps_bills_vote_share <- eda %>% 
  dplyr::select(-StartDate, -year_session) %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("^GRÜNE$", PartyAbbreviation)) %>% 
  left_join(session_leg_index, by = c("SubmissionSession" = "ID")) %>% 
  group_by(PartyAbbreviation, LegislativePeriodNumber) %>% 
  count() %>% 
  left_join(ch_parties_long, by = c("LegislativePeriodNumber", "PartyAbbreviation")) %>%
  dplyr::rename("count_environment_bills" = "n") %>% 
  mutate(bills_per_vote_share = count_environment_bills / VoteShare)

eda_environment_svp_bills_vote_share <- eda %>% 
  dplyr::select(-StartDate, -year_session) %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("^SVP$", PartyAbbreviation)) %>% 
  left_join(session_leg_index, by = c("SubmissionSession" = "ID")) %>% 
  group_by(PartyAbbreviation, LegislativePeriodNumber) %>% 
  count() %>% 
  left_join(ch_parties_long, by = c("LegislativePeriodNumber", "PartyAbbreviation")) %>%
  dplyr::rename("count_environment_bills" = "n") %>% 
  mutate(bills_per_vote_share = count_environment_bills / VoteShare)

eda_environment_mitte_bills_vote_share <- eda %>% 
  dplyr::select(-StartDate, -year_session) %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("^M-E$|^CVP$", PartyAbbreviation)) %>% 
  mutate(PartyAbbreviation = "M-E") %>% 
  left_join(session_leg_index, by = c("SubmissionSession" = "ID")) %>% 
  group_by(PartyAbbreviation, LegislativePeriodNumber) %>% 
  count() %>% 
  left_join(ch_parties_long, by = c("LegislativePeriodNumber", "PartyAbbreviation")) %>%
  dplyr::rename("count_environment_bills" = "n") %>% 
  mutate(bills_per_vote_share = count_environment_bills / VoteShare)

eda_environment_fdp_bills_vote_share <- eda %>% 
  dplyr::select(-StartDate, -year_session) %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("^FDP$", PartyAbbreviation)) %>% 
  left_join(session_leg_index, by = c("SubmissionSession" = "ID")) %>% 
  group_by(PartyAbbreviation, LegislativePeriodNumber) %>% 
  count() %>% 
  left_join(ch_parties_long, by = c("LegislativePeriodNumber", "PartyAbbreviation")) %>%
  dplyr::rename("count_environment_bills" = "n") %>% 
  mutate(bills_per_vote_share = count_environment_bills / VoteShare)

eda_environment_glp_bills_vote_share <- eda %>% 
  dplyr::select(-StartDate, -year_session) %>% 
  dplyr::filter(grepl(set_tag, TagNames)) %>% 
  dplyr::filter(grepl("glp", PartyAbbreviation)) %>% 
  mutate(PartyAbbreviation = "GLP") %>% 
  left_join(session_leg_index, by = c("SubmissionSession" = "ID")) %>% 
  group_by(PartyAbbreviation, LegislativePeriodNumber) %>% 
  count() %>% 
  left_join(ch_parties_long, by = c("LegislativePeriodNumber", "PartyAbbreviation")) %>%
  dplyr::rename("count_environment_bills" = "n") %>% 
  mutate(bills_per_vote_share = count_environment_bills / VoteShare)

eda_plot <- rbind(eda_environment_sp_bills_vote_share,
                  eda_environment_gps_bills_vote_share,
                  eda_environment_svp_bills_vote_share,
                  eda_environment_mitte_bills_vote_share,
                  eda_environment_fdp_bills_vote_share,
                  eda_environment_glp_bills_vote_share
                  )

ggplot(eda_plot, aes(x = LegislativePeriodNumber, y = bills_per_vote_share, color = PartyAbbreviation)) +
  geom_line() + geom_point()

ggplot(eda_environment_sp_bills_vote_share, aes(x = SubmissionSession, y = bills_per_vote_share)) +
  geom_line() + geom_point()

```



## Overall Bill Length

```{r echo=FALSE, message=FALSE, warning=FALSE}

wc_plot <- all_businesses_wc %>% left_join(eda, by = "BusinessShortNumber")

wc_all_bills <- wc_plot %>% dplyr::select(BusinessShortNumber, SubmittedText_clean_wc)

wc_all_sp_bills <- wc_plot %>% dplyr::filter(grepl("SP", PartyAbbreviation))
wc_all_gps_bills <- wc_plot %>% dplyr::filter(grepl("GRÜNE", PartyAbbreviation))

wc_environmental_bills <- wc_plot %>% 
  dplyr::select(BusinessShortNumber, SubmittedText_clean_wc, TagNames) %>% 
  dplyr::filter(grepl("Umwelt", TagNames))

wc_environmental_sp_bills <- wc_plot %>% 
  dplyr::select(BusinessShortNumber, SubmittedText_clean_wc, TagNames, PartyAbbreviation) %>% 
  dplyr::filter(grepl("Umwelt", TagNames)) %>% 
  dplyr::filter(grepl("SP", PartyAbbreviation))

wc_environmental_gps_bills <- wc_plot %>% 
  dplyr::select(BusinessShortNumber, SubmittedText_clean_wc, TagNames, PartyAbbreviation) %>% 
  dplyr::filter(grepl("Umwelt", TagNames)) %>% 
  dplyr::filter(grepl("GRÜNE", PartyAbbreviation))


# Party Plots
all_bills_plot <- ggplot(wc_all_bills, aes(x = SubmittedText_clean_wc)) +
  geom_boxplot() + coord_flip() +
  scale_y_continuous(breaks = NULL) + # remove y-axis labels
  xlab("Word Count") + ggtitle("All Bills") + 
  xlim(0, 1000) +
  theme_minimal()

all_sp_bills_plot <- ggplot(wc_all_sp_bills, aes(x = SubmittedText_clean_wc)) +
  geom_boxplot() + coord_flip() +
  scale_y_continuous(breaks = NULL) + # remove y-axis labels
  xlab("") + ggtitle("All SP Bills") +
  xlim(0, 1000) +
  theme_minimal() +
  theme(axis.text.y = element_blank())

all_gps_bills_plot <- ggplot(wc_all_gps_bills, aes(x = SubmittedText_clean_wc)) +
  geom_boxplot() + coord_flip() +
  scale_y_continuous(breaks = NULL) + # remove y-axis labels
  xlab("") + ggtitle("All Green Bills") +
  xlim(0, 1000) +
  theme_minimal() +
  theme(axis.text.y = element_blank())

grid.arrange(all_bills_plot, all_sp_bills_plot, all_gps_bills_plot, ncol = 3)

# Environmental Plots
environmental_bills_plot <- ggplot(wc_environmental_bills, aes(x = SubmittedText_clean_wc)) +
  geom_boxplot() + coord_flip() +
  scale_y_continuous(breaks = NULL) + # remove y-axis labels
  xlab("Word Count") + ggtitle("All Environmental Bills") +
  xlim(0, 1000) +
  theme_minimal()

environmental_sp_bills_plot <- ggplot(wc_environmental_sp_bills, aes(x = SubmittedText_clean_wc)) +
  geom_boxplot() + coord_flip() +
  scale_y_continuous(breaks = NULL) + # remove y-axis labels
  xlab("") + ggtitle("SP Environmental Bills") +
  xlim(0, 1000) +
  theme_minimal() +
  theme(axis.text.y = element_blank())

environmental_gps_bills_plot <- ggplot(wc_environmental_gps_bills, aes(x = SubmittedText_clean_wc)) +
  geom_boxplot() + coord_flip() +
  scale_y_continuous(breaks = NULL) + # remove y-axis labels
  xlab("") + ggtitle("Green Environmental Bills") +
  xlim(0, 1000) +
  theme_minimal() +
  theme(axis.text.y = element_blank())

grid.arrange(environmental_bills_plot, environmental_sp_bills_plot, environmental_gps_bills_plot, ncol = 3)
```


## Bill Length Over Time

```{r echo=FALSE, message=FALSE, warning=FALSE}
plot_time <- wc_all_bills %>% 
  left_join(dplyr::select(eda, BusinessShortNumber, year_session), by = "BusinessShortNumber") %>% 
  drop_na(year_session) %>% 
  mutate(year_session = year(year_session))

# Standard deviation of the median of the bills word count
sd_bills <- plot_time %>% 
  group_by(year_session) %>% 
  summarise(sd = sd(SubmittedText_clean_wc, na.rm = T))
round(sd(sd_bills$sd), 2)

plot_time_environment <- wc_all_bills %>% 
  left_join(dplyr::select(eda, BusinessShortNumber, year_session, TagNames), by = "BusinessShortNumber") %>% 
  dplyr::filter(grepl("Umwelt", TagNames)) %>% 
  drop_na(year_session) %>% 
  mutate(year_session = year(year_session))

# Standard deviation of the median of the bills word count
sd_bills <- plot_time_environment %>% 
  group_by(year_session) %>% 
  summarise(sd = sd(SubmittedText_clean_wc, na.rm = T))
round(sd(sd_bills$sd), 2)

plot_time_sp <- wc_all_bills %>% 
  left_join(dplyr::select(eda, BusinessShortNumber, year_session, PartyAbbreviation), by = "BusinessShortNumber") %>% 
  dplyr::filter(grepl("SP", PartyAbbreviation)) %>% 
  drop_na(year_session) %>% 
  mutate(year_session = year(year_session))

plot_time_greens <- wc_all_bills %>% 
  left_join(dplyr::select(eda, BusinessShortNumber, year_session, PartyAbbreviation), by = "BusinessShortNumber") %>% 
  dplyr::filter(grepl("GRÜNE", PartyAbbreviation)) %>% 
  drop_na(year_session) %>% 
  mutate(year_session = year(year_session))

plot_time_sp_environment <- wc_all_bills %>% 
  left_join(dplyr::select(eda, BusinessShortNumber, year_session, PartyAbbreviation, TagNames), by = "BusinessShortNumber") %>% 
  dplyr::filter(grepl("SP", PartyAbbreviation)) %>% 
  dplyr::filter(grepl("Umwelt", TagNames)) %>% 
  drop_na(year_session) %>% 
  mutate(year_session = year(year_session))

# Standard deviation of the median of the bills word count
sd_bills <- plot_time_sp_environment %>% 
  group_by(year_session) %>% 
  summarise(sd = sd(SubmittedText_clean_wc))
round(sd(sd_bills$sd), 2)

plot_time_greens_environment <- wc_all_bills %>% 
  left_join(dplyr::select(eda, BusinessShortNumber, year_session, PartyAbbreviation, TagNames), by = "BusinessShortNumber") %>% 
  dplyr::filter(grepl("GRÜNE", PartyAbbreviation)) %>% 
  dplyr::filter(grepl("Umwelt", TagNames)) %>% 
  drop_na(year_session) %>% 
  mutate(year_session = year(year_session))

# Standard deviation of the median of the bills word count
sd_bills <- plot_time_greens_environment %>% 
  group_by(year_session) %>% 
  summarise(sd = sd(SubmittedText_clean_wc, na.rm = T))
round(sd(sd_bills$sd), 2)

# Plots
ggplot(plot_time, aes(x = as.factor(year_session), y = SubmittedText_clean_wc)) +
  geom_boxplot() +
  labs(x = " ", y = "Word Count", title = "All Bills: Word Count (Limited to 900 words)") +
  ylim(0, 900) +
  theme_minimal()

ggplot(plot_time_environment, aes(x = as.factor(year_session), y = SubmittedText_clean_wc)) +
  geom_boxplot() +
  labs(x = " ", y = "Word Count", title = "Environment Bills: Word Count") +
  ylim(0, 900) +
  theme_minimal()

ggplot(plot_time_sp, aes(x = as.factor(year_session), y = SubmittedText_clean_wc)) +
  geom_boxplot() +
  labs(x = " ", y = "Word Count", title = "SP Bills: Word Count") +
  ylim(0, 900) +
  theme_minimal()

ggplot(plot_time_greens, aes(x = as.factor(year_session), y = SubmittedText_clean_wc)) +
  geom_boxplot() +
  labs(x = " ", y = "Word Count", title = "Green Bills: Word Count") +
  ylim(0, 900) +
  theme_minimal()

ggplot(plot_time_sp_environment, aes(x = as.factor(year_session), y = SubmittedText_clean_wc)) +
  geom_boxplot() +
  labs(x = " ", y = "Word Count", title = "SP Environment Bills: Word Count") +
  ylim(0, 900) +
  theme_minimal()

ggplot(plot_time_greens_environment, aes(x = as.factor(year_session), y = SubmittedText_clean_wc)) +
  geom_boxplot() +
  labs(x = " ", y = "Word Count", title = "Green Environment Bills: Word Count") +
  ylim(0, 900) +
  theme_minimal()
```
