==> scripts/05_embedding_eda.Rmd <==
---
title: "Embedding Exploratory Data Analysis"
output:
  html_document:
    toc: true
    toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(tidyr)
library(here)
library(gridExtra)

sp_greens_all_bills_distance_ft_embeddings <- readRDS(file = here("data", "sp_greens_all_bills_distance_ft_embeddings.Rda"))
# sp_greens_all_bills_distance_glove_embeddings <- readRDS(file = here("data", "sp_greens_all_bills_distance_glove_embeddings.Rda"))

sp_greens_environmental_bills_distance_ft_embeddings <- readRDS(file = here("data", "sp_greens_environmental_bills_distance_ft_embeddings.Rda"))
# sp_greens_environmental_bills_distance_glove_embeddings <- readRDS(file = here("data", "sp_greens_environmental_bills_distance_glove_embeddings.Rda"))

load(file = here("data", "all_sessions.Rda"))
```

## Distance between Bills in Vector Space
Lower distance = Bills have more similar content.  
Dotted black vertical lines mark parliamentary elections.  
Dashed red vertical lines mark crises: The Fukushima nuclear accident (2011) and Greta Thunberg's climate strike (2018).

**Average distance between all bills introduced by SP and Greens using pre-trained FastText embeddings**

```{r echo=FALSE, message=FALSE, warning=FALSE}
df_plot <- sp_greens_all_bills_distance_ft_embeddings %>% 
  left_join(dplyr::select(all_sessions, ID, StartDate), by = c("SubmissionSession" = "ID")) %>% 
  mutate(session_year = year(StartDate))

df_plot_yearly_avg <- df_plot %>% 
  group_by(session_year) %>%
  summarise(distance = mean(distance)) %>% 
  mutate(session_year = as.Date(paste0(session_year, "-07-02"))) # Middle of the year

session_year_mapping <- df_plot %>%
  dplyr::select(SubmissionSession, session_year) %>%
  distinct() %>%
  arrange(SubmissionSession)

# Generate labels, leaving blanks for repeated years
unique_labels <- session_year_mapping$session_year
# Keep only the first occurrence of each year, replace others with ""
labels_for_plot <- ifelse(duplicated(unique_labels), "", as.character(unique_labels))

# Make sure SubmissionSession is a factor in the order you want
df_plot$SubmissionSession <- factor(df_plot$SubmissionSession,
                                    levels = unique(session_year_mapping$SubmissionSession))

# Plot
ggplot() +
  geom_line(data = df_plot, aes(x = StartDate, y = distance), group = 1, color = "grey") +
  geom_point(data = df_plot, aes(x = StartDate, y = distance), shape = 21, fill = "white", color = "grey") +
  geom_line(data = df_plot_yearly_avg, aes(x = as.Date(session_year), y = distance)) +
  geom_point(data = df_plot_yearly_avg, aes(x = as.Date(session_year), y = distance)) +
  # scale_x_discrete(labels = labels_for_plot) + 
  ylim(0, 9) +
  # ggtitle("Average Distance between All Bills introduced by SP and Greens") +
  labs(subtitle = "Average Distance over Time (all bills, by session and by year)") +
  xlab("") + ylab("") + 
  # Mark elections
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "5019"), linetype = "dotted", color = "#333333", size = 0.25) +
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4920"), linetype = "dotted", color = "#333333", size = 0.25) +
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4820"), linetype = "dotted", color = "#333333", size = 0.25) +
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4718"), linetype = "dotted", color = "#333333", size = 0.25) +
  # Mark crisis
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4817"), linetype = "dashed", color = "red", size = 0.25) +
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "5014"), linetype = "dashed", color = "red", size = 0.25) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())
```

<!-- **Average Distance between All Bills introduced by SP and Greens: GloVe** -->

```{r eval=FALSE, message=FALSE, warning=FALSE, include=FALSE}
df_plot <- sp_greens_all_bills_distance_glove_embeddings %>% 
  left_join(dplyr::select(all_sessions, ID, StartDate), by = c("SubmissionSession" = "ID")) %>% 
  mutate(session_year = year(StartDate))

session_year_mapping <- df_plot %>%
  dplyr::select(SubmissionSession, session_year) %>%
  distinct() %>%
  arrange(SubmissionSession)

# Generate labels, leaving blanks for repeated years
unique_labels <- session_year_mapping$session_year
# Keep only the first occurrence of each year, replace others with ""
labels_for_plot <- ifelse(duplicated(unique_labels), "", as.character(unique_labels))

# Make sure SubmissionSession is a factor in the order you want
df_plot$SubmissionSession <- factor(df_plot$SubmissionSession,
                                    levels = unique(session_year_mapping$SubmissionSession))

# Plot
ggplot(df_plot, aes(x = SubmissionSession, y = distance)) +
  geom_line(group = 1) +
  geom_point(shape = 21, fill = "white", color = "black") + 
  scale_x_discrete(labels = labels_for_plot) + 
  ylim(0, 15) +
  # ggtitle("Average Distance between All Bills introduced by SP and Greens") +
  labs(subtitle = "Bills grouped by Session") +
  xlab("") + ylab("") + 
  # Mark elections
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "5019"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4920"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4820"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4718"), linetype = "dotted", color = "#333333", size = 0.25) +
  # Mark crisis
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4817"), linetype = "dashed", color = "red", size = 0.25) +
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "5014"), linetype = "dashed", color = "red", size = 0.25) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())
```

**Average distance between environmental bills introduced by SP and Greens using pre-trained FastText embeddings**

```{r echo=FALSE, message=FALSE, warning=FALSE}
df_plot <- sp_greens_environmental_bills_distance_ft_embeddings %>% 
  left_join(dplyr::select(all_sessions, ID, StartDate), by = c("SubmissionSession" = "ID")) %>% 
  mutate(session_year = year(StartDate))

df_plot_yearly_avg <- df_plot %>% 
  group_by(session_year) %>%
  summarise(distance = mean(distance)) %>% 
  mutate(session_year = as.Date(paste0(session_year, "-07-02"))) # Middle of the year

session_year_mapping <- df_plot %>%
  dplyr::select(SubmissionSession, session_year) %>%
  distinct() %>%
  arrange(SubmissionSession)

# Generate labels, leaving blanks for repeated years
unique_labels <- session_year_mapping$session_year
# Keep only the first occurrence of each year, replace others with ""
labels_for_plot <- ifelse(duplicated(unique_labels), "", as.character(unique_labels))

# Make sure SubmissionSession is a factor in the order you want
df_plot$SubmissionSession <- factor(df_plot$SubmissionSession,
                                    levels = unique(session_year_mapping$SubmissionSession))

# Plot
ggplot() +
  geom_line(data = df_plot, aes(x = StartDate, y = distance), group = 1, color = "grey") +
  geom_point(data = df_plot, aes(x = StartDate, y = distance), shape = 21, fill = "white", color = "grey") +
  geom_line(data = df_plot_yearly_avg, aes(x = as.Date(session_year), y = distance)) +
  geom_point(data = df_plot_yearly_avg, aes(x = as.Date(session_year), y = distance)) +
  ylim(0, 15) +
  # ggtitle("Average Distance between Environmental Bills introduced by SP and Greens") +
  labs(subtitle = "Average Distance over Time (environmental bills, by session and by year)") +
  xlab("") + ylab("") + 
  # Mark elections
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "5019"), linetype = "dotted", color = "#333333", size = 0.25) +
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4920"), linetype = "dotted", color = "#333333", size = 0.25) +
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4820"), linetype = "dotted", color = "#333333", size = 0.25) +
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4718"), linetype = "dotted", color = "#333333", size = 0.25) +
  # Mark crisis
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4817"), linetype = "dashed", color = "red", size = 0.25) +
  # geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "5014"), linetype = "dashed", color = "red", size = 0.25) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())
```


<!-- **Average Distance between Environmental Bills introduced by SP and Greens: GloVe** -->

```{r eval=FALSE, message=FALSE, warning=FALSE, include=FALSE}
df_plot <- sp_greens_environmental_bills_distance_glove_embeddings %>% 
  left_join(dplyr::select(all_sessions, ID, StartDate), by = c("SubmissionSession" = "ID")) %>% 
  mutate(session_year = year(StartDate))

session_year_mapping <- df_plot %>%
  dplyr::select(SubmissionSession, session_year) %>%
  distinct() %>%
  arrange(SubmissionSession)

# Generate labels, leaving blanks for repeated years
unique_labels <- session_year_mapping$session_year
# Keep only the first occurrence of each year, replace others with ""
labels_for_plot <- ifelse(duplicated(unique_labels), "", as.character(unique_labels))

# Make sure SubmissionSession is a factor in the order you want
df_plot$SubmissionSession <- factor(df_plot$SubmissionSession,
                                    levels = unique(session_year_mapping$SubmissionSession))

# Plot
ggplot(df_plot, aes(x = SubmissionSession, y = distance)) +
  geom_line(group = 1) +
  geom_point(shape = 21, fill = "white", color = "black") + 
  scale_x_discrete(labels = labels_for_plot) + 
  ylim(0, 15) +
  # ggtitle("Average Distance between Environmental Bills introduced by SP and Greens") +
  labs(subtitle = "Bills grouped by Session") +
  xlab("") + ylab("") + 
  # Mark elections
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "5019"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4920"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4820"), linetype = "dotted", color = "#333333", size = 0.25) +
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4718"), linetype = "dotted", color = "#333333", size = 0.25) +
  # Mark crisis
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "4817"), linetype = "dashed", color = "red", size = 0.25) +
  geom_vline(xintercept = which(levels(df_plot$SubmissionSession) == "5014"), linetype = "dashed", color = "red", size = 0.25) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5),
        panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank())
```
